Acceptance Criteria
Given that I have started the app, I should input my preference lunch date or it would be today date if it is empty
Given that I have made a request to the /ingredients endpoint, I should receive a JSON response of the ingredients in my fridge.
Given that I get the options of my ingredients, I should choose some ingredients to get the recipes.
Given that an ingredient is past its use-by date (inclusive), I should not be able to choose that ingredient.
Given that I have choosen some ingredients, I should request to the /recipes endpoint with ?ingredients=<title-1>,<title-2>,<title-n> as a query parameter.
Given that I get the result of the available recipes, I should see list of the recipes as well the ingredients to make per recipes.